% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parse_names.R
\name{add_names_to_df}
\alias{add_names_to_df}
\title{Split full name into first and last name and add them to data.frame}
\usage{
add_names_to_df(df, full_name, keep = TRUE, ...)
}
\arguments{
\item{df}{dataframe with the data}

\item{full_name}{name of the column containing full name}

\item{keep}{logical indicating if the column containing full name should be kept}

\item{...}{additional params passed to split_full_name function}
}
\description{
Split full name into first and last name and add them to data.frame
}
