% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parse_names.R
\name{decide_unknown_names}
\alias{decide_unknown_names}
\title{Decide unknown names}
\usage{
decide_unknown_names(names, first_names, last_names)
}
\arguments{
\item{names}{Names}

\item{first_names}{Vector of first names}

\item{last_names}{Vector of last names}
}
\description{
Decide unknown names
}
