% Generated by roxygen2 (4.0.2): do not edit by hand
\name{DirectoryPath.S3}
\alias{DirectoryPath.S3}
\title{Class: DirectoryPath.S3}
\usage{
DirectoryPath.S3(.x, path = NA_character_)
}
\arguments{
\item{.x}{\code{\link{ANY}}. An object of an arbitrary class whose class
attribute should be updated so that it becomes an instance of class
\code{DirectoryPath.S3}. Mainly intended for rapid prototyping
purposes}
}
\value{
Instance of class \code{DirectoryPath.S3}.
}
\description{
Directory class (S3).
}
\section{Fields}{

\describe{
\item{\code{path}}{\code{\link{character}}. Directory path.}
}}
\section{Intended use of this class}{


This S3 class, or to be more precise its constructor function, exists mainly
for rapid prototyping purposes.
This is mainly reflected in the fact, that when specifying \code{.x}, this
constructor function will simply update the \code{class} attribute of
whatever object has been provided.

However, it also allows for a more formal OOP-style of rapid
prototyping by offering explicit \emph{class fields} (all arguments except
\code{.x}). Nevertheless, it is probably advisable to switch to an
explicit formal approach such as \emph{S4} and/or \emph{Reference Classes}
once the package or application has reached a certain state of maturity.
}
\examples{
\dontrun{

## TODO: add example

}
}
\author{
Janko Thyson \email{janko.thyson@rappster.de}
}
\references{
\url{http://github.com/rappster/listr}
}
\seealso{
\code{
  	\link[filesystr]{FilePath.S3}
}
}

