% Generated by roxygen2 (4.0.2): do not edit by hand
\name{ExtendedObjectStructure.S3}
\alias{ExtendedObjectStructure.S3}
\title{Class: ExtendedObjectStructure.S3}
\usage{
ExtendedObjectStructure.S3(.x, struc = data.frame())
}
\arguments{
\item{.x}{\code{\link{ANY}}. An object of an arbitrary class whose class
attribute should be updated so that it becomes an instance of class
\code{ExtendedObjectStructure.S3}. Mainly intended for rapid prototyping
purposes}

\item{struc}{Constructor argument. See field \code{.struc}.}
}
\value{
Instance of class \code{ExtendedObjectStructure.S3}.
}
\description{
Class to represent extended object structures (S3).
}
\section{Fields}{

\describe{
\item{\code{.struc}}{\code{\link{data.frame}}. Object structure.}
}}
\section{Intended use of this class}{


This S3 class, or to be more precise its constructor function, exists mainly
for rapid prototyping purposes.
This is mainly reflected in the fact, that when specifying \code{.x}, this
constructor function will simply update the \code{class} attribute of
whatever object has been provided.

However, it also allows for a more formal OOP-style of rapid
prototyping by offering explicit \emph{class fields} (all arguments except
\code{.x}). Nevertheless, it is probably advisable to switch to an
explicit formal approach such as \emph{S4} and/or \emph{Reference Classes}
once the package or application has reached a certain state of maturity.
}
\examples{
\dontrun{

## Informal use (intended mainly for rapid prototyping) //
## Takes *any* object and simply changes the class attributes
ExtendedObjectStructure.S3(data.frame())  
ExtendedObjectStructure.S3("example of other input")  

## Formal use (explicitly using field constructor arguments) //
ExtendedObjectStructure.S3()
ExtendedObjectStructure.S3(struc = getStructure(list(a = 1, b = 2)))

## Recommended: inlcude namespace //
listr::ExtendedObjectStructure.S3(
  struc = getStructure(list(a = 1, b = 2))
)

}
}
\author{
Janko Thyson \email{janko.thyson@rappster.de}
}
\references{
\url{http://github.com/rappster/listr}
}
\seealso{
\code{
    \link[listr]{extendStructure},
    \link[listr]{getRawStructure},
    \link[listr]{getStructure},
    \link[listr]{ObjectStructure.S3}
}
}

