% Generated by roxygen2 (4.0.2): do not edit by hand
\docType{methods}
\name{extendStructure,ObjectStructure.S3-method}
\alias{extendStructure,ObjectStructure.S3-method}
\alias{extendStructure-list-method}
\title{Extend Structure (ObjectStructure.S3)}
\usage{
\S4method{extendStructure}{ObjectStructure.S3}(input, ...)
}
\arguments{
\item{input}{\code{\link{list}}.}

\item{...}{Further arguments to be passed to subsequent functions/methods.}
}
\value{
\code{\link[listr]{ExtendedObjectStructure.S3}}.
   A \code{data.frame} like
   representation of structural information with an additional class
   attribute.
}
\description{
Extend Structure (ObjectStructure.S3)
}
\section{See generic}{

\code{\link[listr]{extendStructure}}
}
\examples{
\dontrun{

##------------------------------------------------------------------------------  
## ObjectStructure.S3
##------------------------------------------------------------------------------  

input <- list(
  x1 = list(x11 = list(x111 = 1, x112 = 1), x12 = list(x121 = 1, x122 = 1)),
  x2 = list(x21 = "x21"),
  x3 = list("x31"),
  x4 = "x4",
  x5 = list(1:3),
  list(list(1, 2), list(3, 4)),
  list(1:3),
  "char 1",
  "char 2",
  letters[1:3],
  c(1,3,5),
  TRUE,
  new.env(),
  data.frame(x = 1:3, y = 1:3)
)
input <- getStructure(input, extended = FALSE)

extendStructure(input)

}
}
\author{
Janko Thyson \email{janko.thyson@rappster.de}
}
\references{
\url{http://github.com/rappster/listr}
}
\seealso{
\code{
   \link[listr]{extendStructure},
   \link[listr]{getStructure},
   \link[listr]{ObjectStructure.S3},
   \link[listr]{ExtendedObjectStructure.S3}
}
}

