% Generated by roxygen2 (4.0.2): do not edit by hand
\name{extendStructure}
\alias{extendStructure}
\title{Extend Structure (generic)}
\usage{
extendStructure(input, ...)
}
\arguments{
\item{input}{\strong{Signature argument}.
Object containing structure information.}

\item{...}{Further arguments to be passed to subsequent functions/methods.}
}
\description{
Extends object structure as returned by \code{\link[listr]{extendStructure}}.
}
\examples{
\dontrun{

##------------------------------------------------------------------------------  
## ObjectStructure.S3
##------------------------------------------------------------------------------  

input <- list(
  x1 = list(x11 = list(x111 = 1, x112 = 1), x12 = list(x121 = 1, x122 = 1)),
  x2 = list(x21 = "x21"),
  x3 = list("x31"),
  x4 = "x4",
  x5 = list(1:3),
  list(list(1, 2), list(3, 4)),
  list(1:3),
  "char 1",
  "char 2",
  letters[1:3],
  c(1,3,5),
  TRUE,
  new.env(),
  data.frame(x = 1:3, y = 1:3)
)
input <- getStructure(input, extended = FALSE)

extendStructure(input)

}
}
\author{
Janko Thyson \email{janko.thyson@rappster.de}
}
\references{
\url{http://github.com/rappster/listr}
}
\seealso{
\code{
  	\link[listr]{extendStructure-ObjectStructure.S3-method},
    \link[listr]{getStructure}
}
}

