% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parse_titles.R
\name{extract_text_before_titles}
\alias{extract_text_before_titles}
\title{Extract text which is before academic titles}
\usage{
extract_text_before_titles(x)
}
\arguments{
\item{x}{String containing name with academic titles (e.g. Karel Novák prof. RNDr.)}
}
\description{
Extract text which is before academic titles
}
