% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parse_titles.R
\name{extract_titles}
\alias{extract_titles}
\title{Extract titles from string}
\usage{
extract_titles(x)
}
\arguments{
\item{x}{String containing name with academic titles}
}
\description{
Extract titles from string
}
