% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parse_names.R
\name{find_last_name}
\alias{find_last_name}
\title{Find last name}
\usage{
find_last_name(x, tolerance)
}
\arguments{
\item{x}{String with a name}

\item{tolerance}{the first and last name are assigned by the probability counted as
probability of a name being a first name = how many times is this name used as first name /
(how many times is this name used as first name + how many times is this name used as last name)
based on data from the Ministry of Interior. If tolerance is 0, only names that are used
only as first names will be labeled as first names etc.}
}
\description{
Find last name
}
