% Generated by roxygen2 (4.0.2): do not edit by hand
\docType{methods}
\name{getByLevel,list-method}
\alias{getByLevel,list-method}
\alias{getByLevel-list-method}
\title{Get By Level (list)}
\usage{
\S4method{getByLevel}{list}(input, level = c(1, 2, 3), parent = FALSE,
  strict = c(0, 1, 2), ...)
}
\arguments{
\item{input}{\code{\link{list}}.}

\item{level}{\code{\link{numeric}} or \code{\link{integer}}.
Level index as returned by \link[listr]{getStructure}.}

\item{strict}{\code{\link{numeric}}.
\itemize{
  \item{\code{0}: } {ignore without warning}
  \item{\code{1}: } {ignore with Warning}
  \item{\code{2}: } {stop with error}
}}

\item{...}{Further arguments to be passed to subsequent functions/methods.}
}
\value{
\code{\link{list}}.
}
\description{
Get By Level (list)
}
\section{See generic}{

\code{\link[listr]{getByLevel}}
}
\examples{
\dontrun{

input <- list(
  x1 = list(x11 = list(x111 = 1, x112 = 1), x12 = list(x121 = 1, x122 = 1)),
  x2 = list(x21 = "x21"),
  x3 = list("x31"),
  x4 = "x4",
  x5 = list(1:3),
  list(list(1, 2), list(3, 4)),
  list(1:3),
  "char 1",
  "char 2",
  letters[1:3],
  c(1,3,5),
  TRUE,
  new.env(),
  data.frame(x = 1:3, y = 1:3)
)

## Type 1 (top-level branches) //
getByType(input, type = 1)
## --> identical to `input`

## Type 2 (intermediate-level branches) //
getByType(input, type = 2)

## Type 3 (leafs) //
getByType(input, type = 3)

}
}
\author{
Janko Thyson \email{janko.thyson@rappster.de}
}
\references{
\url{http://github.com/rappster/listr}
}
\seealso{
\code{
   \link[listr]{getByLevel},
   \link[listr]{getStructure}
}
}

