% Generated by roxygen2 (4.0.2): do not edit by hand
\name{getByLevel}
\alias{getByLevel}
\title{Get By Level (generic)}
\usage{
getByLevel(input, value, outer = FALSE, resolve = FALSE, strict = c(0, 1,
  2), ...)
}
\arguments{
\item{input}{\strong{Signature argument}.
Object containing structure information.}

\item{value}{\code{\link{character}}.
Level as found in return value of \link[listr]{getStructure}.}

\item{outer}{\code{\link{logical}}.
\code{TRUE}: use \code{oindex} instead of \code{index} to retrieve the
  outer element values (i.e., outer list structure is preserved);
\code{FALSE}: use \code{index} to retrieve the inner element values
  (i.e. outer list structure is not preserved).}

\item{resolve}{\code{\link{logical}}.
\code{TRUE}: resolve path to basename component as names for return value;
\code{FALSE}: use path as names for return value.}

\item{strict}{\code{\link{numeric}}.
\itemize{
  \item{\code{0}: } {ignore without warning}
  \item{\code{1}: } {ignore with Warning}
  \item{\code{2}: } {stop with error}
}}

\item{...}{Further arguments to be passed to subsequent functions/methods.}
}
\description{
Retrieves actual values based on their \code{level}.
}
\examples{
\dontrun{

input <- list(
  x1 = list(x11 = list(x111 = list(x1111 = 1, x1112 = 2), x112 = 1), 
    x12 = list(x121 = 1, x122 = 1)),
  x2 = list(x21 = "x21"),
  x3 = list("x31"),
  x4 = "x4",
  x5 = list(1:3),
  list(list(1, 2), list(3, 4)),
  list(1:3),
  "char 1",
  "char 2",
  letters[1:3],
  c(1,3,5),
  TRUE,
  new.env(),
  data.frame(x = 1:3, y = 1:3)
)

## Inspect structure //
(struc <- getStructure(input))

getByLevel(input, value = 1)
getByLevel(input, value = 1, resolve = TRUE)

getByLevel(input, value = 2)
getByLevel(input, value = 2, resolve = TRUE)

getByLevel(input, value = 3)
getByLevel(input, value = 3, resolve = TRUE)

getByLevel(input, value = 4)
getByLevel(input, value = 4, resolve = TRUE)

## Condition handling //
getByLevel(input, value = 5)
try(getByLevel(input, value = 5, strict = 1))
try(getByLevel(input, value = 5, strict = 2))

}
}
\author{
Janko Thyson \email{janko.thyson@rappster.de}
}
\references{
\url{http://github.com/rappster/listr}
}
\seealso{
\code{
  	\link[listr]{getByLevel-list-method},
    \link[listr]{getStructure}
}
}

