% Generated by roxygen2 (4.0.2): do not edit by hand
\docType{methods}
\name{getByName,list-method}
\alias{getByName,list-method}
\alias{getByName-list-method}
\title{Get By name (list)}
\usage{
\S4method{getByName}{list}(input, name, parent = FALSE, strict = c(0, 1, 2),
  ...)
}
\arguments{
\item{input}{\code{\link{list}}.}

\item{name}{\code{\link{numeric}} or \code{\link{integer}}.
Name index as returned by \link[listr]{getStructure}.}

\item{strict}{\code{\link{numeric}}.
\itemize{
  \item{\code{0}: } {ignore without warning}
  \item{\code{1}: } {ignore with Warning}
  \item{\code{2}: } {stop with error}
}}

\item{...}{Further arguments to be passed to subsequent functions/methods.}
}
\value{
\code{\link{list}}.
}
\description{
Get By name (list)
}
\section{See generic}{

\code{\link[listr]{getByName}}
}
\examples{
\dontrun{

## TODO: add example

}
}
\author{
Janko Thyson \email{janko.thyson@rappster.de}
}
\references{
\url{http://github.com/rappster/listr}
}
\seealso{
\code{
   \link[listr]{getByName},
   \link[listr]{getStructure}
}
}

