% Generated by roxygen2 (4.0.2): do not edit by hand
\name{getByName}
\alias{getByName}
\title{Get By Name (generic)}
\usage{
getByName(input, name, parent = FALSE, strict = c(0, 1, 2), ...)
}
\arguments{
\item{input}{\strong{Signature argument}.
Object containing structure information.}

\item{name}{\code{\link{numeric}} or \code{\link{integer}}.
Name index as returned by \link[listr]{getStructure}.}

\item{strict}{\code{\link{numeric}}.
\itemize{
  \item{\code{0}: } {ignore without warning}
  \item{\code{1}: } {ignore with Warning}
  \item{\code{2}: } {stop with error}
}}

\item{...}{Further arguments to be passed to subsequent functions/methods.}
}
\description{
Retrieves actual values based on their \code{name} index.
}
\examples{
\dontrun{

## TODO: add example

}
}
\author{
Janko Thyson \email{janko.thyson@rappster.de}
}
\references{
\url{http://github.com/rappster/listr}
}
\seealso{
\code{
  	\link[listr]{getByName-list-method},
    \link[listr]{getStructure}
}
}

