% Generated by roxygen2 (4.0.2): do not edit by hand
\name{getByPath}
\alias{getByPath}
\title{Get By Path (generic)}
\usage{
getByPath(input, value, outer = FALSE, resolve = FALSE, strict = c(0, 1,
  2), ...)
}
\arguments{
\item{input}{\strong{Signature argument}.
Object containing structure information.}

\item{value}{\code{\link{character}}.
Path as found in return value of \link[listr]{getStructure}.}

\item{outer}{\code{\link{logical}}.
\code{TRUE}: use \code{oindex} instead of \code{index} to retrieve the
  outer element values (i.e., outer list structure is preserved);
\code{FALSE}: use \code{index} to retrieve the inner element values
  (i.e. outer list structure is not preserved).}

\item{resolve}{\code{\link{logical}}.
\code{TRUE}: resolve path to basename component as names for return value;
\code{FALSE}: use path as names for return value.}

\item{strict}{\code{\link{numeric}}.
\itemize{
  \item{\code{0}: } {ignore without warning}
  \item{\code{1}: } {ignore with Warning}
  \item{\code{2}: } {stop with error}
}}

\item{...}{Further arguments to be passed to subsequent functions/methods.}
}
\description{
Retrieves actual values based on their \code{path}.
}
\examples{
\dontrun{

## TODO: add example

}
}
\author{
Janko Thyson \email{janko.thyson@rappster.de}
}
\references{
\url{http://github.com/rappster/listr}
}
\seealso{
\code{
  	\link[listr]{getByPath-list-method},
    \link[listr]{getStructure}
}
}

