% Generated by roxygen2 (4.0.2): do not edit by hand
\docType{methods}
\name{getRawStructure,list-method}
\alias{getRawStructure,list-method}
\alias{getRawStructure-list-method}
\title{Get Raw Structure (list)}
\usage{
\S4method{getRawStructure}{list}(input, ...)
}
\arguments{
\item{input}{\code{\link{list}}.}

\item{...}{Further arguments to be passed to subsequent functions/methods.}
}
\value{
\code{\link[listr]{RawStructure}}. A \code{data.frame} like
   representation of structural information with an additional class
   attribute.
}
\description{
Get Raw Structure (list)
}
\section{See generic}{

\code{\link[listr]{getRawStructure}}
}
\examples{
\dontrun{

## TODO: add example

}
}
\author{
Janko Thyson \email{janko.thyson@rappster.de}
}
\references{
\url{http://github.com/rappster/listr}
}
\seealso{
\code{
   \link[listr]{getRawStructure}
}
}

