% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/list_remove.R
\name{list_remove}
\alias{list_remove}
\title{Remove elements from a list.}
\usage{
list_remove(in_list, ...)
}
\arguments{
\item{in_list}{The list to remove elements from.}

\item{...}{Names or numeric positions of elements to remove.}
}
\value{
The list with the specified elements removed.
}
\description{
Remove elements from a list.
}
\examples{
my_list <- list(a = rnorm(10), b = rnorm(10), c = rnorm(10))
list_remove(my_list, b)
}
