% Generated by roxygen2 (4.0.2): do not edit by hand
\docType{package}
\name{listr}
\alias{listr}
\alias{listr-package}
\title{Package Title}
\description{
Package description
}
\details{
The core functions/methods of this package:
 \itemize{
   \item{\code{\link[listr]{function-name}}: }{

     Description of this function.

     Use empty lines for creating vertical space.
   }
   \item{\code{\link[listr]{function-name}}: }{

     Description of this function.

     Use empty lines for creating vertical space.
   }
}
}
\author{
Janko Thyson \email{janko.thyson@rappster.de}
}
\references{
\url{http://github.com/rappster/listr}
}

