% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parse_gender.R
\name{parse_gender}
\alias{parse_gender}
\title{Parse gender from first and last name columns}
\usage{
parse_gender(df, first_name_col = "first_name",
  last_name_col = "last_names", tolerance = 0.1)
}
\arguments{
\item{df}{Dataframe containing the data}

\item{first_name_col}{Name of the column containing the first names}

\item{last_name_col}{Name of the column containing the last names}

\item{tolerance}{Tolerance for assigning gender based on the frequency of person
with particular name with a given gender}
}
\description{
Parse gender from first and last name columns
}
