% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parse_gender.R
\name{parse_person_gender2}
\alias{parse_person_gender2}
\title{Parse gender of a single person}
\usage{
parse_person_gender2(first_name, last_name, tolerance = 0.1)
}
\arguments{
\item{first_name}{String containing first name}

\item{last_name}{String containing last name}

\item{tolerance}{Tolerance for assigning gender based on the frequency of person
with particular name with a given gender}
}
\description{
Parse gender of a single person
}
