% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parse_titles.R
\name{recode_titles}
\alias{recode_titles}
\title{Extract the highest received academic title}
\usage{
recode_titles(x)
}
\arguments{
\item{x}{String containing titles}
}
\description{
Extract the highest received academic title
}
