% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parse_names.R
\name{split_full_name}
\alias{split_full_name}
\title{Split name}
\usage{
split_full_name(x, reversed = TRUE, tolerance = 0.1)
}
\arguments{
\item{x}{String containing full name}

\item{reversed}{Logical indicating whether the full name is
in reversed order (first name at the end)}

\item{tolerance}{The first and last name are assigned by the probability counted as
probability of a name being a first name = how many times is this name used as first name /
(how many times is this name used as first name + how many times is this name used as last name)
based on data from the Ministry of Interior. If tolerance is 0, only names that are used
only as first names will be labeled as first names etc.}
}
\description{
Split name into first and last name
}
