% Generated by roxygen2 (4.0.2): do not edit by hand
\name{getRawStructure}
\alias{getRawStructure}
\title{Get Raw Structure (generic)}
\usage{
getRawStructure(input, ...)
}
\arguments{
\item{...}{Further arguments to be passed to subsequent functions/methods.}

\item{main}{\strong{Signature argument}.
Object whose structure should be retrieved.}
}
\description{
Retrieves the structural components of an object in a raw form.
}
\examples{
\dontrun{

## TODO: add example

}
}
\author{
Janko Thyson \email{janko.thyson@rappster.de}
}
\references{
\url{http://github.com/rappster/listr}
}
\seealso{
\code{
  	\link[listr]{getRawStructure-list-method}
}
}

